\name{BellMisc-package}
\alias{BellMisc-package}
\alias{BellMisc}
\docType{package}
\title{
Contains utility functions used by Kendon Bell}
\description{
Various utility functions}
\details{
\tabular{ll}{
Package: \tab BellMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-29\cr
License: \tab GPL-3.0\cr
}
Misc functions
}
\author{
Kendon Bell
Maintainer: Kendon Bell <kmb56@berkeley.edu>
}