% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graphPolygonWithCategoriesSouthBrazil}
\alias{graphPolygonWithCategoriesSouthBrazil}
\title{Plot a polygon with discrete colour fill over a polygon with black lines.}
\usage{
graphPolygonWithCategoriesSouthBrazil(colourVar, dataPoly, legendTitle,
  fileName, plotTitle, directoryPlots = getwd(), backPolygon = NULL,
  brewerPalette = "BuGn", fontFam = NULL)
}
\arguments{
\item{colourVar}{categorical variable that defines the gradient
colour. Same length and order as dataPoly.}

\item{dataPoly}{SpatialPolygonsDataFrame to be plotted.}

\item{fileName}{filename (not with directory) as character.}

\item{legendTitle}{legend title as character.}

\item{directoryPlots}{directory to store plots}

\item{backPolygon}{optional background polygon to be
plotted on top of - fortified or raw (prefortifying will be faster if
several maps are being generated).}

\item{brewerPalette}{brewer palette name as character. Default is "BuGn".}
}
\value{
A ggplot object. In addition, a pdf file is written with name \code{fileName}
in directory \code{directoryPlots}.
}
\description{
Plot a polygon with discrete colour fill over a polygon with black lines.
}

