% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPolygonWithGradient.R
\name{graphPolygonWithGradient}
\alias{graphPolygonWithGradient}
\title{Plot a polygon with colour fill over a polygon with black lines.}
\usage{
graphPolygonWithGradient(colourVar, dataPoly, legendTitle, fileName, plotTitle,
  directoryPlots = getwd(), backPolygon = NULL, brewerPalette = "RdYlGn",
  fontFam = NULL, myrev = TRUE, limits = NULL, heightAdj = 1,
  widthAdj = 1)
}
\arguments{
\item{colourVar}{numeric variable that defines the gradient
colour. Same length and order as dataPoly.}

\item{dataPoly}{SpatialPolygonsDataFrame to be plotted.}

\item{legendTitle}{legend title as character.}

\item{fileName}{filename (not with directory or extension) as character.}

\item{directoryPlots}{directory to store plots}

\item{backPolygon}{optional background polygon to be
plotted on top of - fortified or raw (prefortifying will be faster if
several maps are being generated).}

\item{brewerPalette}{brewer palette name as character. Default is "Spectral".}

\item{rev}{logical to indicate if Brewer colours should be reversed.}
}
\value{
a ggplot2 object
}
\description{
Plot a polygon with colour fill over a polygon with black lines.
}

