% Generated by roxygen2 (4.0.2): do not edit by hand
\name{leapsOutofSampleSpatial}
\alias{leapsOutofSampleSpatial}
\title{Linear regression model selection with an out-of-sample criterion for spatial extrapolation.}
\usage{
leapsOutofSampleSpatial(xData, yData, nReps = 100,
  tau = ceiling(NROW(xData)/10), names = NULL, int = TRUE, nbest = 1)
}
\arguments{
\item{xData}{a matrix of predictors}

\item{yData}{a response vector}

\item{nReps}{number of combinations of spatial units to be used in the
estimation}

\item{tau}{number of observations to remove from estimation and use in out-of-sample
prediction. Default is ceiling(x/10).}

\item{int}{Add an intercept to the model (default is TRUE)}

\item{nbest}{number of regressions to report (default is 1)}

\item{regressor}{names}
}
\value{
list of lm objects
}
\description{
Linear regression model selection with an out-of-sample criterion for spatial
extrapolation. Current version uses the minimum mean squared forecast
error method. While leaps reports
subsets for each number of regressors, this reports across all sizes.
Randomly samples several sets of observations to be removed.
Brute force algorithm without much regard for memory conservation.
***Future version should implement a parallel version of this.
}
\details{
Thanks to Thomas Lumley for the similar \code{leaps()} function that selects
with an in-sample criterion. His uses much more efficient algorithms. The
current implementation is only robust for nbest=1.
}

