% Generated by roxygen2 (4.0.1): do not edit by hand
\name{substrRight}
\alias{substrRight}
\title{Substring from the right of a \code{character}}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{\code{character} to be extracted from}

\item{n}{number of characters to extract}
}
\value{
\code{character}
}
\description{
Substring from the right of a \code{character}
}

