% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collapseSpatialPoly.R
\name{collapseSpatialPoly}
\alias{collapseSpatialPoly}
\title{Collapse a SpatialPolygonsDataFrame}
\usage{
collapseSpatialPoly(poly, collapseVarString)
}
\arguments{
\item{poly}{\code{SpatialPolygonsDataFrame} to be collapsed}

\item{collapseVarString}{variable to be collapsed by as a character}
}
\value{
a \code{SpatialPolygonsDataFrame}
}
\description{
Collapse a SpatialPolygonsDataFrame by a variable.
}

