% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertShpPolyToDf.R
\name{convertShpPolyToDf}
\alias{convertShpPolyToDf}
\title{Converts a \code{SpatialPolygonsDataFrame} to a dataframe ()
for plotting in ggplot2.}
\usage{
convertShpPolyToDf(shpPoly)
}
\arguments{
\item{vec}{vector}
}
\value{
\code{numeric}
}
\description{
Converts a \code{SpatialPolygonsDataFrame} to a dataframe ()
for plotting in ggplot2.
}

