% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gdpDeflator}
\alias{gdpDeflator}
\title{Generates a GDP deflator from the World Bank}
\usage{
gdpDeflator(country, start, end)
}
\arguments{
\item{country}{country (ISO-2 character codes, e.g. "BR", "US", "CA")
for which the deflator is needed.}

\item{start}{first year of data}

\item{end}{last year of data}
}
\value{
a data.frame containing the world bank output and a
gdpDeflator index.
}
\description{
Generates a GDP deflator from the World Bank
}

