% Generated by roxygen2 (4.0.2): do not edit by hand
\name{patchKnitrSynctex}
\alias{patchKnitrSynctex}
\title{Patch synctex file for knitr}
\usage{
patchKnitrSynctex(texfile)
}
\arguments{
\item{texfile}{path to base tex file}
}
\value{
None - just some messages.
}
\description{
Fixes up a synctex file so that forward and backward search works
}
\details{
Written by Jan Gleixner and put in this package by Kendon Bell with some minor adjustments.
Jan's notes are: it does use the "blubb-concordance.tex"
file generated by knitr to patch the blubb.synctex[.gz] file generated by latex in order to
fix forward and reverse seach between blubb.pdf and blubb.Rnw
tested only under windows, with one single utf8 Rnw file.
Author: Jan Gleixner <jan.gleixner at gmail>
}

