% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simpleCap}
\alias{simpleCap}
\title{Takes a \code{character} and makes the first letter of each word upper case.}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{\code{character} to be capitalized}
}
\value{
\code{character} with capitalized first letter.
}
\description{
Takes a \code{character} and makes the first letter of each word upper case.
Taken from \link{http://stat.ethz.ch/R-manual/R-patched/library/base/html/chartr.html}
}

