% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whichClosest}
\alias{whichClosest}
\title{Find index(ices) of closest value(s).}
\usage{
whichClosest(x, y)
}
\arguments{
\item{x}{vector to search through}

\item{y}{number to find closed value to}
}
\value{
index(ices) of closest value(s).
}
\description{
Find index(ices) of closest value(s).
}

